Abstracted the animation logic from player
