+++ game-client/game-client/Player.cpp
@@ -31,12 +31,9 @@ Player::Player(sf::Vector2f position, ResourceContainer& resourceContainer)
  */
 Player::Player(sf::Vector2f position, TextureManager& textureManager, b2World& physicsWorld)
 	:Entity(),
-	mTextureManager(textureManager),
-	mCurrentAnimation(nullptr),
-	mAccumulator(0.0f),
-	currentFrame(0)
+	Animatable(textureManager)
 {
-	initAnimationData();
+	loadAnimations();
 	setAnimation(IDLE);
 
 	sf::IntRect& frame = mCurrentAnimation->frames[currentFrame]; 
@@ -53,48 +50,18 @@ Player::Player(sf::Vector2f position, TextureManager& textureManager, b2World& p
 
 Player::~Player()
 {
-	for (auto& data : mAnimations)
-		mTextureManager.deleteTexture(data.second.textureKey);
 }
 
 /**
  * @brief Loads and creates animation data
  * 
  */
-void Player::initAnimationData()
+void Player::loadAnimations()
 {
 	loadAnimation(FALL, 1, 0.0f);
 	loadAnimation(IDLE, 11, 50.0f);
 	loadAnimation(RUN, 12, 50.0f);
 	loadAnimation(JUMP, 1, 0);
-}
-
-/**
- * @brief Creates an AnimationData struct and adds it animations map
- * 
- * @param name filename of the animation png
- * @param frameCount number of frames in png
- * @param frameRate frame display rate for accumulator
- */
-void Player::loadAnimation(const std::string& name, int frameCount, float frameRate)
-{
-	std::vector<sf::IntRect> frames;
-	for (int i = 0; i < frameCount; ++i)
-	{
-		sf::IntRect frame(32 * i, 0, 32, 32);
-		frames.push_back(frame);
-	}
-	mTextureManager.loadTexture(name);
-	
-	AnimationData animData = {
-		name,
-		frames,
-		name,
-		frameCount,
-		frameRate,
-	};
-	
-	mAnimations.emplace(name, animData);
 }
 
 /**
@@ -185,32 +152,6 @@ void Player::update(sf::Time dt)
 	syncPositions();
 }
 
-/**
- * @brief Retrieves the current animation playing
- * 
- * @return AnimationData& reference to the animation currently playing
- */
-AnimationData& Player::currentAnimation()
-{
-	if (!mCurrentAnimation)
-		throw std::runtime_error("Player does not have animation loaded");
-
-	return *mCurrentAnimation;
-}
-
-/**
- * @brief Changes the current animation being played
- * 
- * @param name filename of the animation to play
- */
-void Player::setAnimation(const std::string& name)
-{
-	if (mAnimations.count(name) == 0)
-		throw std::runtime_error("Animation does not exist in Player");
-	
-
-	mCurrentAnimation = &mAnimations.at(name);
-}
 
 /**
  * @brief mIsInteracting getter
