+++ game-client/game-client/game-client.vcxproj
@@ -182,6 +182,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Animatable.cpp" />
     <ClCompile Include="Background.cpp" />
     <ClCompile Include="Entity.cpp" />
     <ClCompile Include="EntityFactory.cpp" />
@@ -195,6 +196,7 @@
     <ClCompile Include="PlayerView.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Animatable.h" />
     <ClInclude Include="Background.h" />
     <ClInclude Include="Entity.h" />
     <ClInclude Include="EntityFactory.h" />
