+++ game-client/game-client/game-client.vcxproj.filters
@@ -60,6 +60,9 @@
     <ClCompile Include="PlayerView.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Animatable.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Program.h">
@@ -90,6 +93,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="PlayerView.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Animatable.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
